refactoring 3-1
