+++ src/Components/Admin/UserDataTable/UserDataTable.js
@@ -1,20 +1,8 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
-import { ADMIN } from "Utils/constants";
-import AuthSelector from "Components/Admin/UserDataTable/AuthSelector";
+import UserDataLine from "./UserDataLine";
 
 const UserDataTable = ({ filteredUsers, handleAuthUpdate }) => {
-  const {
-    authTitle: { choice, admin, teacher, parents },
-  } = ADMIN;
-  const authTitle = [choice, admin, teacher, parents];
-  const [choicedAuth, setChoicedAuth] = useState(-1);
-
-  const handleOptionChange = (auth) => {
-    if (auth === -1) return;
-    setChoicedAuth(auth);
-  };
-
   return (
     <Wrapper>
       <div className="dataTable">
@@ -30,36 +18,9 @@ const UserDataTable = ({ filteredUsers, handleAuthUpdate }) => {
               <th>신용 카드</th>
               <th colSpan="3">권한</th>
             </tr>
-            //분리권장 => comment
-            {filteredUsers?.map(
-              ({ id, name, email, pw, address, dateOfBirth, creditCardNum, authority }) => (
-                <tr key={id}>
-                  <td>{id}</td>
-                  <td>{name}</td>
-                  <td>{email}</td>
-                  <td>{pw}</td>
-                  <td>{address}</td>
-                  <td>{dateOfBirth}</td>
-                  <td>{creditCardNum}</td>
-                  <td>{authTitle.slice(1)[authority]}</td>
-                  <td>
-                    <AuthSelector handleOptionChange={handleOptionChange} authTitle={authTitle} />
-                  </td>
-                  <td>
-                    <button type="button" onClick={() => handleAuthUpdate(id, choicedAuth)}>
-                      수정
-                    </button>
-                  </td>
-                </tr>
-              )
-            )}
+            <UserDataLine {...{ filteredUsers, handleAuthUpdate }} />
           </tbody>
         </table>
-        {filteredUsers.length === 0 && (
-          <EmptyData>
-            <span>일치하는 데이터가 없습니다. 🤢</span>
-          </EmptyData>
-        )}
       </div>
     </Wrapper>
   );
@@ -91,10 +52,4 @@ const Wrapper = styled.div`
   button {
     border: 1px solid ${({ theme }) => theme.color.borderline};
   }
-`;
-
-const EmptyData = styled.div`
-  ${({ theme }) => theme.flexSet()};
-  height: 300px;
-  border: 1px solid ${({ theme }) => theme.color.borderline};
 `;
